refactor(topbar): extract sticky nav logic to hook
